+++ tasks/schedule.sh
@@ -24,9 +24,9 @@ schedule_task() {
 
   echo -e "${YELLOW}Scheduling task $task_name, with ID: $task_id (Duration: $duration minutes, Datetime: $datetime)...${RESET}"
 
-  # Skip scheduling if the task name contains "Google-kalenterin tapahtuma"
-  if [[ "$task_name" == *"Google-kalenterin tapahtuma"* ]]; then
-    echo -e "${YELLOW}Skipping scheduling task: $task_name (ID: $task_id)${RESET}"
+  # Check if task has a label with name "Google-kalenterin tapahtuma"
+  if echo "$task_data" | jq -r '.labels[]' | grep -q "Google-kalenterin tapahtuma"; then
+    echo -e "${YELLOW}Skipping postponing task, because it has the calendar label: $task_name (ID: $task_id)${RESET}"
     return 0
   fi